Implement CURL interface
